Replace explicit tests crate with single implicit integration tests crate
