+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "codegen", "macro", "tests"]
+members = [".", "codegen", "macro"]
 
 [package]
 name = "substrate-subxt"
@@ -47,3 +47,15 @@ sp-runtime = { package = "sp-runtime", git = "https://github.com/paritytech/subs
 sp-version = { package = "sp-version", git = "https://github.com/paritytech/substrate/", branch = "master" }
 
 frame-metadata = "14.0.0"
+
+[dev-dependencies]
+assert_matches = "1.5.0"
+async-std = { version = "1.9.0", features = ["attributes", "tokio1"] }
+env_logger = "0.8.3"
+tempdir = "0.3.7"
+wabt = "0.10.0"
+which = "4.0.2"
+
+sp-arithmetic = { package = "sp-arithmetic", git = "https://github.com/paritytech/substrate/", branch = "master" }
+sp-keyring = { package = "sp-keyring", git = "https://github.com/paritytech/substrate/", branch = "master" }
+
